pyproject.toml
[tool.poetry]
name = "python-postgres"
version = "0.1.0"
description = ""
authors = ["shuyao <shuyao.guo@tu-berlin.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
pandas = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
